refactor(kanban): extract claim/lock lifecycle (kanban_db.py god-file slice R3)
